change name children

+++ database/migrations/2022_06_21_131325_create_immunizations_table.php
@@ -14,7 +14,7 @@ class CreateImmunizationsTable extends Migration
     public function up()
     {
         Schema::create('immunizations', function (Blueprint $table) {
-            $table->unsignedInteger('child_id');
+            $table->unsignedBigInteger('child_id');
             $table->unsignedBigInteger('toi_id');
             $table->foreign('child_id')->references('id')->on('children')->onDelete('cascade');
             $table->foreign('toi_id')->references('id')->on('type_of_immunizations')->onDelete('cascade');
